requirements.txt
selenium == 4.23.1
pytest == 8.3.2
openpyxl == 3.1.5
allure-pytest == 2.13.5